requirements.txt
certifi==2019.6.16
dataclasses==0.6
Django==2.2.4
djangorestframework==3.10.2
djongo==1.2.33
dnspython==1.16.0
pymongo==3.8.0
pytz==2019.2
sqlparse==0.2.4
wincertstore==0.2
